Update vchan-socket-proxy, fix QMP reconnect race
Vchan server may not notice when client quickly disconnects and
reconnects. In this case, vchan-socket-proxy would not reconnect the
target UNIX socket (to Qemu in this case). This means, a new vchan
client will not get QMP handshake and will timeout.
Fix this by updating vchan-socket-proxy to a version with in-band
handshake that allows detecting such case and handling it appropriately.
It requires enabling it explicitly with --reconnect-marker=1 ("1" is not
"on", it's a 0x01 byte value) - matching to similar option on libxl
side.

Fixes QubesOS/qubes-issues#6824